move new tests to existing distill step
